refactor(orphan-pass): apply post-merge cleanup findings
- Hoist `mock_pg_tx` fixture from `tests/unit/test_entity_store.py` to `tests/unit/conftest.py` so the transactional-store mock is reusable across any test module that exercises `acquire() + conn.transaction()`. Replaces a fixture that nearly duplicated `test_entity_dedup.py`'s local `mock_pg`.
- Drop the dead `startswith("DELETE")` guard in `delete_identity_by_library_name`: asyncpg's `Connection.execute` returns the status tag verbatim ("DELETE N") for a `DELETE` statement; the conditional was never reachable.
- Hoist the per-test `from scripts.entity_resolution.__main__ import (...)` blocks in the integration test file to a single module-level import. Five identical inline imports collapse to one.
- Tighten the caller log message in `prune_orphan_identities` when `merge_identity_by_library_name` returns False — the previous "(no row to merge)" wording missed the idempotent self-merge case that the docstring already covers.

No behavior change.
